+++ src/plain_merge_sort.cpp
@@ -10,16 +10,16 @@
 #include "sort.h"
 
 namespace PlainMergeSort{
-    void Solution::_mergeArray(vector<int> &nums, vector<int> &buffer, int left, int mid, int right)
+    void Solution::_mergeArray(vector<int> &nums, vector<int> &buffer, size_t left, size_t mid, size_t right)
     {
         if (left >= mid || mid >= right)
         {
             return;
         }
 
-        int i = left;
-        int j = mid;
-        int k = left;
+        size_t i = left;
+        size_t j = mid;
+        size_t k = left;
 
         while (i < mid && j < right)
         {
@@ -43,20 +43,20 @@ namespace PlainMergeSort{
             buffer[k++] = nums[j++];
         }
 
-        for (int index = left; index < k; ++index)
+        for (size_t index = left; index < k; ++index)
         {
             nums[index] = buffer[index];
         }
     }
 
-    void Solution::_sortArray(vector<int> &nums, vector<int> &buffer, int left, int right)
+    void Solution::_sortArray(vector<int> &nums, vector<int> &buffer, size_t left, size_t right)
     {
         if (right - left <= 1)
         {
             return;
         }
 
-        int mid = left + ((right - left) >> 1);
+        size_t mid = left + ((right - left) >> 1);
 
         _sortArray(nums, buffer, left, mid);
         _sortArray(nums, buffer, mid, right);
